+++ test/twitter.test.ts
@@ -1,11 +1,14 @@
 import { test, expect } from "vitest"
-import { getFollowerCount } from "../src"
+import { getBrowserContext, getFollowerCount, destroyBrowser } from "../src"
 
 test("twitter", async () => {
   const count = await getFollowerCount({
     type: "twitter",
     username: "cristiano",
+    browserContext: await getBrowserContext(),
   })
+
+  await destroyBrowser()
 
   console.log(count)
   expect(count).toBeGreaterThan(0)